full set of name changes in test file
